Added package.json file with initial contents.
# Aider chat conversation:

USER: lets add a package.json

ASSISTANT: To add a `package.json` file to your project, you can use the following `SEARCH/REPLACE block`. Since this is a new file, the `SEARCH` section will be empty, and the `REPLACE` section will contain the initial contents of the `package.json` file.

```json
aider/frontend/src/package.json
<<<<<<< SEARCH

=======
{
  "name": "frontend",
  "version": "1.0.0",
  "description": "Frontend for the chat application",
  "main": "index.js",
  "scripts": {
    "start": "react-scripts start",
    "build": "react-scripts build",
    "test": "react-scripts test",
    "eject": "react-scripts eject"
  },
  "dependencies": {
    "axios": "^0.21.1",
    "react": "^17.0.2",
    "react-dom": "^17.0.2",
    "react-scripts": "4.0.3"
  },
  "browserslist": {
    "production": [
      ">0.2%",
      "not dead",
      "not op_mini all"
    ],
    "development": [
      "last 1 chrome version",
      "last 1 firefox version",
      "last 1 safari version"
    ]
  },
  "eslintConfig": {
    "extends": [
      "react-app",
      "react-app/jest"
    ]
  },
  "private": true
}
>>>>>>> REPLACE
```

Please note that the dependencies listed in the `package.json` file are just an example and should be adjusted to match the actual dependencies of your project. The versions of the packages should also be updated to the latest stable versions as needed.